feat: allow optional secret config
Allow a 'secret' config to be read as a layer _after_ env and _before_ 'local'. The 'secret' config is at an absolute path. It is supposed to be provided by the environment to allow sensitive config to be used from sources which are not files packaged with the application, such as externally managed secrets.